requirements.txt
alembic==1.14.0
APScheduler==3.11.0
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-DotEnv==0.1.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==1.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
msgspec @ git+https://github.com/jcrist/msgspec.git@595c33c4a71c6d0c539b82233982a65819e240cf
openssh-wrapper==0.4
paramiko==3.5.0
proxmoxer==2.1.0
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.2.1
